refactor(validator): modularize persistence layer into focused modules
* Break down monolithic simple_persistence.rs (2857 lines) into 14 specialized modules
* Create types.rs for shared persistence types (NodeStats, CapacityEntry, MinerData, etc.)
* Extract verification_logs.rs with 6 verification-related methods
* Extract rentals.rs with rental management and query operations
* Extract verification.rs for verification scheduling
* Create 6 profile modules (hardware, speedtest, network, docker, nat, storage)
* Extend miner_nodes.rs with 13 node-related query methods
* Extend miners.rs with 8 miner management methods
* Reduce simple_persistence.rs to core initialization and trait implementation (505 lines)
* Update mod.rs to declare and export new modules
* Preserve all SQL logic, error handling, and business rules unchanged
* Maintain 100% test pass rate (5/5 tests passing)
