+++ crates/basilica-validator/src/persistence/simple_persistence.rs
@@ -1,42 +1,16 @@
-use chrono::{DateTime, Utc};
-use serde_json::Value;
-use sqlx::{QueryBuilder, Row, SqlitePool};
-use tracing::{info, warn};
-use uuid::Uuid;
+use sqlx::SqlitePool;
+use tracing::info;
 
-use crate::persistence::entities::{Rental, RentalStatus, VerificationLog};
+use crate::persistence::types::RentalFilter;
 use crate::persistence::ValidatorPersistence;
 use crate::rental::{RentalInfo, RentalState};
 
-/// Extract GPU memory size in GB from GPU name string
-fn extract_gpu_memory_gb(gpu_name: &str) -> u32 {
-    use regex::Regex;
-
-    let re = Regex::new(r"(\d+)GB").unwrap();
-    if let Some(captures) = re.captures(gpu_name) {
-        captures[1].parse().unwrap_or(0)
-    } else {
-        0
-    }
-}
-
-/// Filter criteria for querying rentals
-#[derive(Default)]
-pub struct RentalFilter {
-    pub rental_id: Option<String>,
-    pub validator_hotkey: Option<String>,
-    pub exclude_states: Option<Vec<RentalState>>,
-    pub order_by_created_desc: bool,
-}
-
-/// Simplified persistence implementation for quick testing
 #[derive(Debug, Clone)]
 pub struct SimplePersistence {
-    pool: SqlitePool,
+    pub(crate) pool: SqlitePool,
 }
 
 impl SimplePersistence {
-    /// Get access to the underlying database pool
     pub fn pool(&self) -> &SqlitePool {
         &self.pool
     }
@@ -113,7 +87,9 @@ impl SimplePersistence {
     }
 
     async fn initialize_collateral_scan_status(&self) -> Result<(), anyhow::Error> {
+        use chrono::Utc;
         use collateral_contract::config::CONTRACT_DEPLOYED_BLOCK_NUMBER;
+        use tracing::warn;
 
         let now = Utc::now().to_rfc3339();
         let insert_query = r#"
@@ -135,2218 +111,6 @@ impl SimplePersistence {
         }
 
         Ok(())
-    }
-
-    pub async fn create_verification_log(
-        &self,
-        log: &VerificationLog,
-    ) -> Result<(), anyhow::Error> {
-        let query = r#"
-            INSERT INTO verification_logs (
-                id, node_id, validator_hotkey, verification_type, timestamp,
-                score, success, details, duration_ms, error_message, created_at, updated_at
-            ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-        "#;
-
-        sqlx::query(query)
-            .bind(log.id.to_string())
-            .bind(&log.node_id)
-            .bind(&log.validator_hotkey)
-            .bind(&log.verification_type)
-            .bind(log.timestamp.to_rfc3339())
-            .bind(log.score)
-            .bind(if log.success { 1 } else { 0 })
-            .bind(&serde_json::to_string(&log.details)?)
-            .bind(log.duration_ms)
-            .bind(&log.error_message)
-            .bind(log.created_at.to_rfc3339())
-            .bind(log.updated_at.to_rfc3339())
-            .execute(&self.pool)
-            .await?;
-
-        tracing::info!(
-            verification_id = %log.id,
-            node_id = %log.node_id,
-            success = %log.success,
-            score = %log.score,
-            "Verification log created"
-        );
-
-        Ok(())
-    }
-
-    /// Query verification logs with filtering and pagination
-    pub async fn query_verification_logs(
-        &self,
-        node_id: Option<&str>,
-        success_only: Option<bool>,
-        limit: u32,
-        offset: u32,
-    ) -> Result<Vec<VerificationLog>, anyhow::Error> {
-        let mut query = String::from(
-            "SELECT id, node_id, validator_hotkey, verification_type, timestamp,
-             score, success, details, duration_ms, error_message, created_at, updated_at
-             FROM verification_logs WHERE 1=1",
-        );
-
-        let mut conditions = Vec::new();
-
-        if let Some(exec_id) = node_id {
-            conditions.push(format!("node_id = '{exec_id}'"));
-        }
-
-        if let Some(success) = success_only {
-            conditions.push(format!("success = {}", if success { 1 } else { 0 }));
-        }
-
-        if !conditions.is_empty() {
-            query.push_str(" AND ");
-            query.push_str(&conditions.join(" AND "));
-        }
-
-        query.push_str(" ORDER BY timestamp DESC LIMIT ? OFFSET ?");
-
-        let rows = sqlx::query(&query)
-            .bind(limit as i64)
-            .bind(offset as i64)
-            .fetch_all(&self.pool)
-            .await?;
-
-        let mut logs = Vec::new();
-        for row in rows {
-            logs.push(self.row_to_verification_log(row)?);
-        }
-
-        Ok(logs)
-    }
-
-    /// Get node statistics from verification logs
-    pub async fn get_node_stats(&self, node_id: &str) -> Result<Option<NodeStats>, anyhow::Error> {
-        let row = sqlx::query(
-            "SELECT
-                COUNT(*) as total_verifications,
-                SUM(CASE WHEN success = 1 THEN 1 ELSE 0 END) as successful_verifications,
-                AVG(score) as avg_score,
-                AVG(duration_ms) as avg_duration_ms,
-                MIN(timestamp) as first_verification,
-                MAX(timestamp) as last_verification
-             FROM verification_logs
-             WHERE node_id = ?",
-        )
-        .bind(node_id)
-        .fetch_optional(&self.pool)
-        .await?;
-
-        if let Some(row) = row {
-            let total: i64 = row.get("total_verifications");
-            if total == 0 {
-                return Ok(None);
-            }
-
-            let stats = NodeStats {
-                node_id: node_id.to_string(),
-                total_verifications: total as u64,
-                successful_verifications: row.get::<i64, _>("successful_verifications") as u64,
-                average_score: row.get("avg_score"),
-                average_duration_ms: row.get("avg_duration_ms"),
-                first_verification: row.get::<Option<String>, _>("first_verification").map(|s| {
-                    DateTime::parse_from_rfc3339(&s)
-                        .unwrap()
-                        .with_timezone(&Utc)
-                }),
-                last_verification: row.get::<Option<String>, _>("last_verification").map(|s| {
-                    DateTime::parse_from_rfc3339(&s)
-                        .unwrap()
-                        .with_timezone(&Utc)
-                }),
-            };
-
-            Ok(Some(stats))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Get available capacity based on successful verifications
-    pub async fn get_available_capacity(
-        &self,
-        min_score: Option<f64>,
-        min_success_rate: Option<f64>,
-        limit: u32,
-        offset: u32,
-    ) -> Result<Vec<CapacityEntry>, anyhow::Error> {
-        let min_score = min_score.unwrap_or(0.0);
-        let min_success_rate = min_success_rate.unwrap_or(0.0);
-
-        let rows = sqlx::query(
-            "SELECT
-                node_id,
-                COUNT(*) as total_verifications,
-                SUM(CASE WHEN success = 1 THEN 1 ELSE 0 END) as successful_verifications,
-                AVG(score) as avg_score,
-                MAX(timestamp) as last_verification,
-                MAX(details) as latest_details
-             FROM verification_logs
-             GROUP BY node_id
-             HAVING avg_score >= ?
-                AND (CAST(successful_verifications AS REAL) / CAST(total_verifications AS REAL)) >= ?
-             ORDER BY avg_score DESC, last_verification DESC
-             LIMIT ? OFFSET ?"
-        )
-        .bind(min_score)
-        .bind(min_success_rate)
-        .bind(limit as i64)
-        .bind(offset as i64)
-        .fetch_all(&self.pool)
-        .await?;
-
-        let mut entries = Vec::new();
-        for row in rows {
-            let node_id: String = row.get("node_id");
-            let total_verifications: i64 = row.get("total_verifications");
-            let successful_verifications: i64 = row.get("successful_verifications");
-            let avg_score: f64 = row.get("avg_score");
-            let last_verification: String = row.get("last_verification");
-            let latest_details: String = row.get("latest_details");
-
-            let success_rate = if total_verifications > 0 {
-                successful_verifications as f64 / total_verifications as f64
-            } else {
-                0.0
-            };
-
-            let details: Value = serde_json::from_str(&latest_details).unwrap_or(Value::Null);
-
-            entries.push(CapacityEntry {
-                node_id,
-                verification_score: avg_score,
-                success_rate,
-                last_verification: DateTime::parse_from_rfc3339(&last_verification)
-                    .unwrap()
-                    .with_timezone(&Utc),
-                hardware_info: details,
-                total_verifications: total_verifications as u64,
-            });
-        }
-
-        Ok(entries)
-    }
-
-    /// Get available nodes for rental (not currently rented)
-    pub async fn get_available_nodes(
-        &self,
-        min_gpu_memory: Option<u32>,
-        gpu_type: Option<String>,
-        min_gpu_count: Option<u32>,
-        location: Option<basilica_common::LocationProfile>,
-    ) -> Result<Vec<AvailableNodeData>, anyhow::Error> {
-        // Build the base query with LEFT JOIN to find nodes without active rentals
-        // Also join with gpu_uuid_assignments to get actual GPU data
-        // And join with hardware profile to get CPU/RAM information
-        // And join with network profile to get location information
-        // And join with speedtest profile to get network speed information
-        let mut query_str = String::from(
-            "SELECT
-                me.node_id,
-                me.miner_id,
-                me.status,
-                me.gpu_count,
-                m.verification_score,
-                m.uptime_percentage,
-                GROUP_CONCAT(gua.gpu_name) as gpu_names,
-                ehp.cpu_model,
-                ehp.cpu_cores,
-                ehp.ram_gb,
-                enp.city,
-                enp.region,
-                enp.country,
-                esp.download_mbps,
-                esp.upload_mbps,
-                esp.test_timestamp
-            FROM miner_nodes me
-            JOIN miners m ON me.miner_id = m.id
-            LEFT JOIN rentals r ON me.node_id = r.node_id
-                AND r.miner_id = me.miner_id
-                AND r.state IN ('Active', 'Provisioning', 'active', 'provisioning')
-            LEFT JOIN gpu_uuid_assignments gua ON me.node_id = gua.node_id AND gua.miner_id = me.miner_id
-            LEFT JOIN node_hardware_profile ehp ON me.node_id = ehp.node_id AND me.miner_id = 'miner_' || ehp.miner_uid
-            LEFT JOIN node_network_profile enp ON me.node_id = enp.node_id AND me.miner_id = 'miner_' || enp.miner_uid
-            LEFT JOIN node_speedtest_profile esp ON me.node_id = esp.node_id AND me.miner_id = 'miner_' || esp.miner_uid
-            WHERE r.id IS NULL
-                AND (me.status IS NULL OR me.status != 'offline')",
-        );
-
-        // Add location filters if specified (case-insensitive comparison)
-        if let Some(ref loc) = location {
-            if let Some(ref country) = loc.country {
-                query_str.push_str(&format!(" AND LOWER(enp.country) = LOWER('{}')", country));
-            }
-            if let Some(ref region) = loc.region {
-                query_str.push_str(&format!(" AND LOWER(enp.region) = LOWER('{}')", region));
-            }
-            if let Some(ref city) = loc.city {
-                query_str.push_str(&format!(" AND LOWER(enp.city) = LOWER('{}')", city));
-            }
-        }
-
-        query_str.push_str(" GROUP BY me.node_id");
-
-        // Add GPU count filter if specified (use HAVING since we're grouping)
-        if let Some(min_count) = min_gpu_count {
-            query_str.push_str(&format!(" HAVING COUNT(gua.gpu_uuid) >= {}", min_count));
-        }
-
-        let rows = sqlx::query(&query_str).fetch_all(&self.pool).await?;
-
-        let mut nodes = Vec::new();
-        for row in rows {
-            // Get GPU data from gpu_uuid_assignments join
-            let gpu_names: Option<String> = row.get("gpu_names");
-
-            // Parse GPU specs from gpu_uuid_assignments data only
-            let mut gpu_specs: Vec<crate::api::types::GpuSpec> = vec![];
-
-            if let Some(names) = gpu_names {
-                if !names.is_empty() {
-                    // Parse GPU names from GROUP_CONCAT result
-                    for gpu_name in names.split(',') {
-                        // Extract memory from GPU name
-                        let memory_gb = extract_gpu_memory_gb(gpu_name);
-
-                        gpu_specs.push(crate::api::types::GpuSpec {
-                            name: gpu_name.to_string(),
-                            memory_gb,
-                            compute_capability: "8.0".to_string(), // Default, could be parsed from prover results
-                        });
-                    }
-                }
-            }
-
-            // Apply GPU memory filter if specified
-            if let Some(min_memory) = min_gpu_memory {
-                let meets_memory = gpu_specs.iter().any(|gpu| gpu.memory_gb >= min_memory);
-                if !meets_memory && !gpu_specs.is_empty() {
-                    continue;
-                }
-            }
-
-            // Apply GPU type filter if specified
-            if let Some(ref gpu_type_filter) = gpu_type {
-                let matches_type = gpu_specs.iter().any(|gpu| {
-                    gpu.name
-                        .to_lowercase()
-                        .contains(&gpu_type_filter.to_lowercase())
-                });
-                if !matches_type && !gpu_specs.is_empty() {
-                    continue;
-                }
-            }
-
-            // Get hardware profile data if available, otherwise use defaults
-            let cpu_model: Option<String> = row.get("cpu_model");
-            let cpu_cores: Option<i32> = row.get("cpu_cores");
-            let ram_gb: Option<i32> = row.get("ram_gb");
-
-            let cpu_specs = crate::api::types::CpuSpec {
-                cores: cpu_cores.unwrap_or(0) as u32,
-                model: cpu_model.unwrap_or_else(|| "Unknown".to_string()),
-                memory_gb: ram_gb.unwrap_or(0) as u32,
-            };
-
-            // Get network profile data for location
-            let city: Option<String> = row.get("city");
-            let region: Option<String> = row.get("region");
-            let country: Option<String> = row.get("country");
-
-            // Always use LocationProfile for consistent formatting
-            let location_profile = basilica_common::LocationProfile::new(city, region, country);
-            let location = Some(location_profile.to_string());
-
-            // Get speed test data if available
-            let download_mbps: Option<f64> = row.get("download_mbps");
-            let upload_mbps: Option<f64> = row.get("upload_mbps");
-            let test_timestamp_str: Option<String> = row.get("test_timestamp");
-
-            let speed_test_timestamp = test_timestamp_str.and_then(|ts| {
-                chrono::DateTime::parse_from_rfc3339(&ts)
-                    .ok()
-                    .map(|dt| dt.with_timezone(&chrono::Utc))
-            });
-
-            nodes.push(AvailableNodeData {
-                node_id: row.get("node_id"),
-                miner_id: row.get("miner_id"),
-                gpu_specs,
-                cpu_specs,
-                location,
-                verification_score: row.get("verification_score"),
-                uptime_percentage: row.get("uptime_percentage"),
-                status: row.get("status"),
-                download_mbps,
-                upload_mbps,
-                speed_test_timestamp,
-            });
-        }
-
-        Ok(nodes)
-    }
-
-    /// Helper function to convert database row to VerificationLog
-    fn row_to_verification_log(
-        &self,
-        row: sqlx::sqlite::SqliteRow,
-    ) -> Result<VerificationLog, anyhow::Error> {
-        let id_str: String = row.get("id");
-        let details_str: String = row.get("details");
-        let timestamp_str: String = row.get("timestamp");
-        let created_at_str: String = row.get("created_at");
-        let updated_at_str: String = row.get("updated_at");
-
-        Ok(VerificationLog {
-            id: Uuid::parse_str(&id_str)?,
-            node_id: row.get("node_id"),
-            validator_hotkey: row.get("validator_hotkey"),
-            verification_type: row.get("verification_type"),
-            timestamp: DateTime::parse_from_rfc3339(&timestamp_str)?.with_timezone(&Utc),
-            score: row.get("score"),
-            success: row.get::<i64, _>("success") == 1,
-            details: serde_json::from_str(&details_str)?,
-            duration_ms: row.get("duration_ms"),
-            error_message: row.get("error_message"),
-            created_at: DateTime::parse_from_rfc3339(&created_at_str)?.with_timezone(&Utc),
-            updated_at: DateTime::parse_from_rfc3339(&updated_at_str)?.with_timezone(&Utc),
-        })
-    }
-
-    /// Create a new rental record
-    pub async fn create_rental(&self, rental: &Rental) -> Result<(), anyhow::Error> {
-        let query = r#"
-            INSERT INTO rentals (
-                id, node_id, customer_public_key, docker_image, env_vars,
-                gpu_requirements, ssh_access_info, max_duration_hours, cost_per_hour,
-                status, created_at, updated_at, started_at, terminated_at,
-                termination_reason, total_cost
-            ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-        "#;
-
-        let status_str = match rental.status {
-            RentalStatus::Pending => "Pending",
-            RentalStatus::Active => "Active",
-            RentalStatus::Terminated => "Terminated",
-            RentalStatus::Failed => "Failed",
-        };
-
-        sqlx::query(query)
-            .bind(rental.id.to_string())
-            .bind(&rental.node_id)
-            .bind(&rental.customer_public_key)
-            .bind(&rental.docker_image)
-            .bind(
-                rental
-                    .env_vars
-                    .as_ref()
-                    .map(|v| serde_json::to_string(v).unwrap()),
-            )
-            .bind(serde_json::to_string(&rental.gpu_requirements)?)
-            .bind(serde_json::to_string(&rental.ssh_access_info)?)
-            .bind(rental.max_duration_hours as i64)
-            .bind(rental.cost_per_hour)
-            .bind(status_str)
-            .bind(rental.created_at.to_rfc3339())
-            .bind(rental.updated_at.to_rfc3339())
-            .bind(rental.started_at.map(|dt| dt.to_rfc3339()))
-            .bind(rental.terminated_at.map(|dt| dt.to_rfc3339()))
-            .bind(&rental.termination_reason)
-            .bind(rental.total_cost)
-            .execute(&self.pool)
-            .await?;
-
-        tracing::info!(
-            rental_id = %rental.id,
-            node_id = %rental.node_id,
-            status = ?rental.status,
-            "Rental created"
-        );
-
-        Ok(())
-    }
-
-    /// Get rental by ID
-    pub async fn get_rental(&self, rental_id: &Uuid) -> Result<Option<Rental>, anyhow::Error> {
-        let row = sqlx::query("SELECT * FROM rentals WHERE id = ?")
-            .bind(rental_id.to_string())
-            .fetch_optional(&self.pool)
-            .await?;
-
-        if let Some(row) = row {
-            Ok(Some(self.row_to_rental(row)?))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Update rental record
-    pub async fn update_rental(&self, rental: &Rental) -> Result<(), anyhow::Error> {
-        let status_str = match rental.status {
-            RentalStatus::Pending => "Pending",
-            RentalStatus::Active => "Active",
-            RentalStatus::Terminated => "Terminated",
-            RentalStatus::Failed => "Failed",
-        };
-
-        let query = r#"
-            UPDATE rentals SET
-                status = ?, updated_at = ?, started_at = ?,
-                terminated_at = ?, termination_reason = ?, total_cost = ?
-            WHERE id = ?
-        "#;
-
-        sqlx::query(query)
-            .bind(status_str)
-            .bind(rental.updated_at.to_rfc3339())
-            .bind(rental.started_at.map(|dt| dt.to_rfc3339()))
-            .bind(rental.terminated_at.map(|dt| dt.to_rfc3339()))
-            .bind(&rental.termination_reason)
-            .bind(rental.total_cost)
-            .bind(rental.id.to_string())
-            .execute(&self.pool)
-            .await?;
-
-        tracing::info!(
-            rental_id = %rental.id,
-            status = ?rental.status,
-            "Rental updated"
-        );
-
-        Ok(())
-    }
-
-    /// Check if an node has an active rental
-    pub async fn has_active_rental(
-        &self,
-        node_id: &str,
-        miner_id: &str,
-    ) -> Result<bool, anyhow::Error> {
-        let query = r#"
-            SELECT COUNT(*) as count
-            FROM rentals
-            WHERE node_id = ?
-                AND miner_id = ?
-                AND state = 'active'
-        "#;
-
-        let row = sqlx::query(query)
-            .bind(node_id)
-            .bind(miner_id)
-            .fetch_one(&self.pool)
-            .await?;
-
-        let count: i64 = row.get("count");
-        Ok(count > 0)
-    }
-
-    /// Helper function to parse rental state from string
-    fn parse_rental_state(state_str: &str, rental_id: &str) -> RentalState {
-        match state_str {
-            "provisioning" => RentalState::Provisioning,
-            "active" => RentalState::Active,
-            "stopping" => RentalState::Stopping,
-            "stopped" => RentalState::Stopped,
-            "failed" => RentalState::Failed,
-            unknown => {
-                warn!(
-                    "Unknown rental state '{}' for rental {}, defaulting to Failed",
-                    unknown, rental_id
-                );
-                RentalState::Failed
-            }
-        }
-    }
-
-    /// Helper function to parse a rental row from the database
-    fn parse_rental_row(
-        &self,
-        row: sqlx::sqlite::SqliteRow,
-        node_details: crate::api::types::NodeDetails,
-    ) -> Result<RentalInfo, anyhow::Error> {
-        let state_str: String = row.get("state");
-        let created_at_str: String = row.get("created_at");
-        let container_spec_str: String = row.get("container_spec");
-        let rental_id: String = row.get("id");
-        let node_id: String = row.get("node_id");
-        let metadata: String = row.get("metadata");
-
-        // Use existing parse_rental_state for consistency
-        let state = Self::parse_rental_state(&state_str, &rental_id);
-
-        Ok(RentalInfo {
-            rental_id,
-            validator_hotkey: row.get("validator_hotkey"),
-            node_id,
-            container_id: row.get("container_id"),
-            ssh_session_id: row.get("ssh_session_id"),
-            ssh_credentials: row.get("ssh_credentials"),
-            state,
-            created_at: DateTime::parse_from_rfc3339(&created_at_str)?.with_timezone(&Utc),
-            container_spec: serde_json::from_str(&container_spec_str)?,
-            miner_id: row.get::<String, _>("miner_id"),
-            node_details,
-            end_user_ssh_credentials: row.get("end_user_ssh_credentials"),
-            metadata: serde_json::from_str(&metadata)?,
-        })
-    }
-
-    /// Query rentals with flexible filtering criteria
-    async fn query_rentals(&self, filter: RentalFilter) -> Result<Vec<RentalInfo>, anyhow::Error> {
-        let mut builder = QueryBuilder::new("SELECT * FROM rentals");
-        let mut has_where = false;
-
-        // Build WHERE clause dynamically
-        if let Some(rental_id) = filter.rental_id {
-            builder.push(" WHERE id = ");
-            builder.push_bind(rental_id);
-            has_where = true;
-        }
-
-        if let Some(validator_hotkey) = filter.validator_hotkey {
-            builder.push(if has_where { " AND " } else { " WHERE " });
-            builder.push("validator_hotkey = ");
-            builder.push_bind(validator_hotkey);
-            has_where = true;
-        }
-
-        if let Some(exclude_states) = filter.exclude_states {
-            if !exclude_states.is_empty() {
-                builder.push(if has_where { " AND " } else { " WHERE " });
-                builder.push("state NOT IN (");
-                for (i, state) in exclude_states.iter().enumerate() {
-                    if i > 0 {
-                        builder.push(", ");
-                    }
-                    // IMPORTANT: Convert to lowercase for database
-                    builder.push_bind(match state {
-                        RentalState::Provisioning => "provisioning",
-                        RentalState::Active => "active",
-                        RentalState::Stopping => "stopping",
-                        RentalState::Stopped => "stopped",
-                        RentalState::Failed => "failed",
-                    });
-                }
-                builder.push(")");
-            }
-        }
-
-        if filter.order_by_created_desc {
-            builder.push(" ORDER BY created_at DESC");
-        }
-
-        let query = builder.build();
-        let rows = query.fetch_all(&self.pool).await?;
-
-        // Parse all rows and fetch node details for each
-        let mut rentals = Vec::new();
-        for row in rows {
-            // Extract node_id and miner_id from the row first
-            let node_id: String = row.get("node_id");
-            let miner_id: String = row.get("miner_id");
-
-            // Fetch node details or use defaults if not found
-            let node_details = match self.get_node_details(&node_id, &miner_id).await {
-                Ok(Some(details)) => details,
-                _ => {
-                    // Default node details if not found
-                    crate::api::types::NodeDetails {
-                        id: node_id.clone(),
-                        gpu_specs: vec![],
-                        cpu_specs: crate::api::types::CpuSpec {
-                            cores: 0,
-                            model: "Unknown".to_string(),
-                            memory_gb: 0,
-                        },
-                        location: None,
-                        network_speed: None,
-                    }
-                }
-            };
-
-            rentals.push(self.parse_rental_row(row, node_details)?);
-        }
-
-        Ok(rentals)
-    }
-
-    /// Helper function to convert database row to Rental
-    fn row_to_rental(&self, row: sqlx::sqlite::SqliteRow) -> Result<Rental, anyhow::Error> {
-        let id_str: String = row.get("id");
-        let env_vars_str: Option<String> = row.get("env_vars");
-        let gpu_requirements_str: String = row.get("gpu_requirements");
-        let ssh_access_info_str: String = row.get("ssh_access_info");
-        let status_str: String = row.get("status");
-        let created_at_str: String = row.get("created_at");
-        let updated_at_str: String = row.get("updated_at");
-        let started_at_str: Option<String> = row.get("started_at");
-        let terminated_at_str: Option<String> = row.get("terminated_at");
-
-        let status = match status_str.as_str() {
-            "Pending" => RentalStatus::Pending,
-            "Active" => RentalStatus::Active,
-            "Terminated" => RentalStatus::Terminated,
-            "Failed" => RentalStatus::Failed,
-            _ => return Err(anyhow::anyhow!("Invalid rental status: {}", status_str)),
-        };
-
-        Ok(Rental {
-            id: Uuid::parse_str(&id_str)?,
-            node_id: row.get("node_id"),
-            customer_public_key: row.get("customer_public_key"),
-            docker_image: row.get("docker_image"),
-            env_vars: env_vars_str.map(|s| serde_json::from_str(&s)).transpose()?,
-            gpu_requirements: serde_json::from_str(&gpu_requirements_str)?,
-            ssh_access_info: serde_json::from_str(&ssh_access_info_str)?,
-            max_duration_hours: row.get::<i64, _>("max_duration_hours") as u32,
-            cost_per_hour: row.get("cost_per_hour"),
-            status,
-            created_at: DateTime::parse_from_rfc3339(&created_at_str)?.with_timezone(&Utc),
-            updated_at: DateTime::parse_from_rfc3339(&updated_at_str)?.with_timezone(&Utc),
-            started_at: started_at_str.map(|s| {
-                DateTime::parse_from_rfc3339(&s)
-                    .unwrap()
-                    .with_timezone(&Utc)
-            }),
-            terminated_at: terminated_at_str.map(|s| {
-                DateTime::parse_from_rfc3339(&s)
-                    .unwrap()
-                    .with_timezone(&Utc)
-            }),
-            termination_reason: row.get("termination_reason"),
-            total_cost: row.get("total_cost"),
-        })
-    }
-
-    /// Get all registered miners
-    pub async fn get_all_registered_miners(&self) -> Result<Vec<MinerData>, anyhow::Error> {
-        self.get_registered_miners(0, 10000).await
-    }
-
-    /// Get registered miners with pagination
-    pub async fn get_registered_miners(
-        &self,
-        offset: u32,
-        page_size: u32,
-    ) -> Result<Vec<MinerData>, anyhow::Error> {
-        let rows = sqlx::query(
-            "SELECT
-                id, hotkey, endpoint, verification_score, uptime_percentage,
-                last_seen, registered_at, node_info,
-                (SELECT COUNT(*) FROM miner_nodes WHERE miner_id = miners.id) as node_count
-             FROM miners
-             ORDER BY registered_at DESC
-             LIMIT ? OFFSET ?",
-        )
-        .bind(page_size as i64)
-        .bind(offset as i64)
-        .fetch_all(&self.pool)
-        .await?;
-
-        let mut miners = Vec::new();
-        for row in rows {
-            let node_info_str: String = row.get("node_info");
-            let node_count: i64 = row.get("node_count");
-            let last_seen_str: String = row.get("last_seen");
-            let registered_at_str: String = row.get("registered_at");
-
-            miners.push(MinerData {
-                miner_id: row.get("id"),
-                hotkey: row.get("hotkey"),
-                endpoint: row.get("endpoint"),
-                node_count: node_count as u32,
-                verification_score: row.get("verification_score"),
-                uptime_percentage: row.get("uptime_percentage"),
-                last_seen: chrono::NaiveDateTime::parse_from_str(
-                    &last_seen_str,
-                    "%Y-%m-%d %H:%M:%S",
-                )
-                .map(|dt| DateTime::from_naive_utc_and_offset(dt, Utc))
-                .or_else(|_| {
-                    DateTime::parse_from_rfc3339(&last_seen_str).map(|dt| dt.with_timezone(&Utc))
-                })?,
-                registered_at: chrono::NaiveDateTime::parse_from_str(
-                    &registered_at_str,
-                    "%Y-%m-%d %H:%M:%S",
-                )
-                .map(|dt| DateTime::from_naive_utc_and_offset(dt, Utc))
-                .or_else(|_| {
-                    DateTime::parse_from_rfc3339(&registered_at_str)
-                        .map(|dt| dt.with_timezone(&Utc))
-                })?,
-                node_info: serde_json::from_str(&node_info_str)
-                    .unwrap_or(Value::Object(serde_json::Map::new())),
-            });
-        }
-
-        Ok(miners)
-    }
-
-    /// Register a new miner
-    pub async fn register_miner(
-        &self,
-        miner_id: &str,
-        hotkey: &str,
-        endpoint: &str,
-        nodes: &[crate::api::types::NodeRegistration],
-    ) -> Result<(), anyhow::Error> {
-        let now = Utc::now().to_rfc3339();
-        let node_info = serde_json::to_string(&nodes)?;
-
-        let mut tx = self.pool.begin().await?;
-
-        // Validate that ssh_endpoint are not already registered
-        for node in nodes {
-            let existing =
-                sqlx::query("SELECT COUNT(*) as count FROM miner_nodes WHERE ssh_endpoint = ?")
-                    .bind(&node.ssh_endpoint)
-                    .fetch_one(&mut *tx)
-                    .await?;
-
-            let count: i64 = existing.get("count");
-            if count > 0 {
-                return Err(anyhow::anyhow!(
-                    "Node with ssh_endpoint {} is already registered",
-                    node.ssh_endpoint
-                ));
-            }
-        }
-
-        sqlx::query(
-            "INSERT INTO miners (id, hotkey, endpoint, last_seen, registered_at, updated_at, node_info)
-             VALUES (?, ?, ?, ?, ?, ?, ?)"
-        )
-        .bind(miner_id)
-        .bind(hotkey)
-        .bind(endpoint)
-        .bind(&now)
-        .bind(&now)
-        .bind(&now)
-        .bind(&node_info)
-        .execute(&mut *tx)
-        .await?;
-
-        for node in nodes {
-            let node_id = Uuid::new_v4().to_string();
-
-            sqlx::query(
-                "INSERT INTO miner_nodes (id, miner_id, node_id, ssh_endpoint, gpu_count, created_at, updated_at)
-                 VALUES (?, ?, ?, ?, ?, ?, ?)"
-            )
-            .bind(&node_id)
-            .bind(miner_id)
-            .bind(&node.node_id)
-            .bind(&node.ssh_endpoint)
-            .bind(node.gpu_count as i64)
-            .bind(&now)
-            .bind(&now)
-            .execute(&mut *tx)
-            .await?;
-        }
-
-        tx.commit().await?;
-        Ok(())
-    }
-
-    /// Get miner by ID
-    pub async fn get_miner_by_id(
-        &self,
-        miner_id: &str,
-    ) -> Result<Option<MinerData>, anyhow::Error> {
-        let row = sqlx::query(
-            "SELECT
-                id, hotkey, endpoint, verification_score, uptime_percentage,
-                last_seen, registered_at, node_info,
-                (SELECT COUNT(*) FROM miner_nodes WHERE miner_id = miners.id) as node_count
-             FROM miners
-             WHERE id = ?",
-        )
-        .bind(miner_id)
-        .fetch_optional(&self.pool)
-        .await?;
-
-        if let Some(row) = row {
-            let node_info_str: String = row.get("node_info");
-            let node_count: i64 = row.get("node_count");
-            let last_seen_str: String = row.get("last_seen");
-            let registered_at_str: String = row.get("registered_at");
-
-            Ok(Some(MinerData {
-                miner_id: row.get("id"),
-                hotkey: row.get("hotkey"),
-                endpoint: row.get("endpoint"),
-                node_count: node_count as u32,
-                verification_score: row.get("verification_score"),
-                uptime_percentage: row.get("uptime_percentage"),
-                last_seen: chrono::NaiveDateTime::parse_from_str(
-                    &last_seen_str,
-                    "%Y-%m-%d %H:%M:%S",
-                )
-                .map(|dt| DateTime::from_naive_utc_and_offset(dt, Utc))
-                .or_else(|_| {
-                    DateTime::parse_from_rfc3339(&last_seen_str).map(|dt| dt.with_timezone(&Utc))
-                })?,
-                registered_at: chrono::NaiveDateTime::parse_from_str(
-                    &registered_at_str,
-                    "%Y-%m-%d %H:%M:%S",
-                )
-                .map(|dt| DateTime::from_naive_utc_and_offset(dt, Utc))
-                .or_else(|_| {
-                    DateTime::parse_from_rfc3339(&registered_at_str)
-                        .map(|dt| dt.with_timezone(&Utc))
-                })?,
-                node_info: serde_json::from_str(&node_info_str)
-                    .unwrap_or(Value::Object(serde_json::Map::new())),
-            }))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Update miner information
-    pub async fn update_miner(
-        &self,
-        miner_id: &str,
-        request: &crate::api::types::UpdateMinerRequest,
-    ) -> Result<(), anyhow::Error> {
-        let now = Utc::now().to_rfc3339();
-
-        if let Some(endpoint) = &request.endpoint {
-            let result = sqlx::query("UPDATE miners SET endpoint = ?, updated_at = ? WHERE id = ?")
-                .bind(endpoint)
-                .bind(&now)
-                .bind(miner_id)
-                .execute(&self.pool)
-                .await?;
-
-            if result.rows_affected() == 0 {
-                return Err(anyhow::anyhow!("Miner not found"));
-            }
-        }
-
-        if let Some(nodes) = &request.nodes {
-            // When updating nodes, we need to handle the miner_nodes table
-            let mut tx = self.pool.begin().await?;
-
-            // First, validate that new ssh_endpoints aren't already registered by other miners
-            for node in nodes {
-                let existing = sqlx::query(
-                    "SELECT COUNT(*) as count FROM miner_nodes
-                     WHERE ssh_endpoint = ? AND miner_id != ?",
-                )
-                .bind(&node.ssh_endpoint)
-                .bind(miner_id)
-                .fetch_one(&mut *tx)
-                .await?;
-
-                let count: i64 = existing.get("count");
-                if count > 0 {
-                    return Err(anyhow::anyhow!(
-                        "Node with ssh_endpoint {} is already registered by another miner",
-                        node.ssh_endpoint
-                    ));
-                }
-            }
-
-            // Delete existing nodes for this miner
-            sqlx::query("DELETE FROM miner_nodes WHERE miner_id = ?")
-                .bind(miner_id)
-                .execute(&mut *tx)
-                .await?;
-
-            // Insert new nodes
-            for node in nodes {
-                let node_id = Uuid::new_v4().to_string();
-
-                sqlx::query(
-                    "INSERT INTO miner_nodes (id, miner_id, node_id, ssh_endpoint, gpu_count, created_at, updated_at)
-                     VALUES (?, ?, ?, ?, ?, ?, ?)"
-                )
-                .bind(&node_id)
-                .bind(miner_id)
-                .bind(&node.node_id)
-                .bind(&node.ssh_endpoint)
-                .bind(node.gpu_count as i64)
-                .bind(&now)
-                .bind(&now)
-                .execute(&mut *tx)
-                .await?;
-            }
-
-            // Also update the node_info JSON in the miners table
-            let node_info = serde_json::to_string(nodes)?;
-            let result =
-                sqlx::query("UPDATE miners SET node_info = ?, updated_at = ? WHERE id = ?")
-                    .bind(&node_info)
-                    .bind(&now)
-                    .bind(miner_id)
-                    .execute(&mut *tx)
-                    .await?;
-
-            if result.rows_affected() == 0 {
-                tx.rollback().await?;
-                return Err(anyhow::anyhow!("Miner not found"));
-            }
-
-            tx.commit().await?;
-        }
-
-        Ok(())
-    }
-
-    /// Remove miner
-    pub async fn remove_miner(&self, miner_id: &str) -> Result<(), anyhow::Error> {
-        let result = sqlx::query("DELETE FROM miners WHERE id = ?")
-            .bind(miner_id)
-            .execute(&self.pool)
-            .await?;
-
-        if result.rows_affected() == 0 {
-            Err(anyhow::anyhow!("Miner not found"))
-        } else {
-            Ok(())
-        }
-    }
-
-    /// Get miner health status
-    pub async fn get_miner_health(
-        &self,
-        miner_id: &str,
-    ) -> Result<Option<MinerHealthData>, anyhow::Error> {
-        let rows = sqlx::query(
-            "SELECT node_id, status, last_health_check, created_at
-             FROM miner_nodes
-             WHERE miner_id = ?",
-        )
-        .bind(miner_id)
-        .fetch_all(&self.pool)
-        .await?;
-
-        if rows.is_empty() {
-            return Ok(None);
-        }
-
-        let mut node_health = Vec::new();
-        let mut latest_check = Utc::now() - chrono::Duration::hours(24);
-
-        for row in rows {
-            let last_health_str: Option<String> = row.get("last_health_check");
-            let created_at_str: String = row.get("created_at");
-
-            let last_seen = if let Some(health_str) = last_health_str {
-                DateTime::parse_from_rfc3339(&health_str)?.with_timezone(&Utc)
-            } else {
-                DateTime::parse_from_rfc3339(&created_at_str)?.with_timezone(&Utc)
-            };
-
-            if last_seen > latest_check {
-                latest_check = last_seen;
-            }
-
-            node_health.push(NodeHealthData {
-                node_id: row.get("node_id"),
-                status: row
-                    .get::<Option<String>, _>("status")
-                    .unwrap_or_else(|| "unknown".to_string()),
-                last_seen,
-            });
-        }
-
-        Ok(Some(MinerHealthData {
-            last_health_check: latest_check,
-            node_health,
-        }))
-    }
-
-    /// Schedule verification for miner
-    pub async fn schedule_verification(
-        &self,
-        miner_id: &str,
-        verification_id: &str,
-        verification_type: &str,
-        node_id: Option<&str>,
-    ) -> Result<(), anyhow::Error> {
-        let now = Utc::now().to_rfc3339();
-
-        sqlx::query(
-            "INSERT INTO verification_requests (id, miner_id, verification_type, node_id, scheduled_at, created_at)
-             VALUES (?, ?, ?, ?, ?, ?)"
-        )
-        .bind(verification_id)
-        .bind(miner_id)
-        .bind(verification_type)
-        .bind(node_id)
-        .bind(&now)
-        .bind(&now)
-        .execute(&self.pool)
-        .await?;
-
-        Ok(())
-    }
-
-    /// Get miner nodes
-    pub async fn get_miner_nodes(&self, miner_id: &str) -> Result<Vec<NodeData>, anyhow::Error> {
-        let rows = sqlx::query(
-            "SELECT
-                me.node_id,
-                GROUP_CONCAT(gua.gpu_name) as gpu_names,
-                ehp.cpu_model,
-                ehp.cpu_cores,
-                ehp.ram_gb,
-                enp.city,
-                enp.region,
-                enp.country
-             FROM miner_nodes me
-             LEFT JOIN gpu_uuid_assignments gua ON me.node_id = gua.node_id AND gua.miner_id = me.miner_id
-             LEFT JOIN node_hardware_profile ehp ON me.node_id = ehp.node_id AND me.miner_id = 'miner_' || ehp.miner_uid
-             LEFT JOIN node_network_profile enp ON me.node_id = enp.node_id AND me.miner_id = 'miner_' || enp.miner_uid
-             WHERE me.miner_id = ?
-             GROUP BY me.node_id,
-                      ehp.cpu_model, ehp.cpu_cores, ehp.ram_gb,
-                      enp.city, enp.region, enp.country",
-        )
-        .bind(miner_id)
-        .fetch_all(&self.pool)
-        .await?;
-
-        let mut nodes = Vec::new();
-        for row in rows {
-            // Get GPU data from gpu_uuid_assignments join
-            let gpu_names: Option<String> = row.get("gpu_names");
-
-            // Parse GPU specs from gpu_uuid_assignments data
-            let mut gpu_specs: Vec<crate::api::types::GpuSpec> = vec![];
-
-            if let Some(names) = gpu_names {
-                if !names.is_empty() {
-                    // Parse GPU names from GROUP_CONCAT result
-                    for gpu_name in names.split(',') {
-                        // Extract memory from GPU name
-                        let memory_gb = extract_gpu_memory_gb(gpu_name);
-
-                        gpu_specs.push(crate::api::types::GpuSpec {
-                            name: gpu_name.to_string(),
-                            memory_gb,
-                            compute_capability: "8.0".to_string(),
-                        });
-                    }
-                }
-            }
-
-            // Get hardware profile data from node_hardware_profile table
-            let cpu_model: Option<String> = row.get("cpu_model");
-            let cpu_cores: Option<i32> = row.get("cpu_cores");
-            let ram_gb: Option<i32> = row.get("ram_gb");
-
-            let cpu_specs = crate::api::types::CpuSpec {
-                cores: cpu_cores.unwrap_or(0) as u32,
-                model: cpu_model.unwrap_or_else(|| "Unknown".to_string()),
-                memory_gb: ram_gb.unwrap_or(0) as u32,
-            };
-
-            // Get network profile data for location
-            let city: Option<String> = row.get("city");
-            let region: Option<String> = row.get("region");
-            let country: Option<String> = row.get("country");
-
-            // Always use LocationProfile for consistent formatting
-            let location_profile = basilica_common::LocationProfile::new(city, region, country);
-            let location = Some(location_profile.to_string());
-
-            nodes.push(NodeData {
-                node_id: row.get("node_id"),
-                gpu_specs,
-                cpu_specs,
-                location,
-            });
-        }
-
-        Ok(nodes)
-    }
-
-    /// Get miner ID by node ID
-    pub async fn get_miner_id_by_node(&self, node_id: &str) -> Result<String, anyhow::Error> {
-        let miner_id: String = sqlx::query(
-            "SELECT miner_id FROM miner_nodes \
-                 WHERE node_id = ? \
-                 LIMIT 1",
-        )
-        .bind(node_id)
-        .fetch_one(&self.pool)
-        .await?
-        .get("miner_id");
-
-        Ok(miner_id)
-    }
-
-    /// get node ssh-endpoint by node ID, return None if not found
-    pub async fn get_node_ssh_endpoint(
-        &self,
-        node_id: &str,
-        miner_id: &str,
-    ) -> Result<Option<String>, anyhow::Error> {
-        let row = sqlx::query(
-            "SELECT ssh_endpoint FROM miner_nodes \
-                 WHERE node_id = ? AND miner_id = ? \
-                 LIMIT 1",
-        )
-        .bind(node_id)
-        .bind(miner_id)
-        .fetch_optional(&self.pool)
-        .await?;
-
-        Ok(row.map(|r| r.get("ssh_endpoint")))
-    }
-
-    /// Get detailed node information including GPU and CPU specs
-    pub async fn get_node_details(
-        &self,
-        node_id: &str,
-        miner_id: &str,
-    ) -> Result<Option<crate::api::types::NodeDetails>, anyhow::Error> {
-        // Get node info with GPU data, hardware profile, network profile, and speed test data
-        let row = sqlx::query(
-            "SELECT
-                me.node_id,
-                GROUP_CONCAT(gua.gpu_name) as gpu_names,
-                ehp.cpu_model,
-                ehp.cpu_cores,
-                ehp.ram_gb,
-                enp.city,
-                enp.region,
-                enp.country,
-                esp.download_mbps,
-                esp.upload_mbps,
-                esp.test_timestamp
-             FROM miner_nodes me
-             LEFT JOIN gpu_uuid_assignments gua ON me.node_id = gua.node_id AND gua.miner_id = me.miner_id
-             LEFT JOIN node_hardware_profile ehp ON me.node_id = ehp.node_id AND me.miner_id = 'miner_' || ehp.miner_uid
-             LEFT JOIN node_network_profile enp ON me.node_id = enp.node_id AND me.miner_id = 'miner_' || enp.miner_uid
-             LEFT JOIN node_speedtest_profile esp ON me.node_id = esp.node_id AND me.miner_id = 'miner_' || esp.miner_uid
-             WHERE me.node_id = ? AND me.miner_id = ?
-             GROUP BY me.node_id,
-                      ehp.cpu_model, ehp.cpu_cores, ehp.ram_gb,
-                      enp.city, enp.region, enp.country,
-                      esp.download_mbps, esp.upload_mbps, esp.test_timestamp
-             LIMIT 1",
-        )
-        .bind(node_id)
-        .bind(miner_id)
-        .fetch_optional(&self.pool)
-        .await?;
-
-        if let Some(row) = row {
-            let node_id: String = row.get("node_id");
-
-            // Get GPU data from gpu_uuid_assignments join
-            let gpu_names: Option<String> = row.get("gpu_names");
-
-            // Parse GPU specs from gpu_uuid_assignments data
-            let mut gpu_specs: Vec<crate::api::types::GpuSpec> = vec![];
-
-            if let Some(names) = gpu_names {
-                if !names.is_empty() {
-                    // Parse GPU names from GROUP_CONCAT result
-                    for gpu_name in names.split(',') {
-                        // Extract memory from GPU name
-                        let memory_gb = extract_gpu_memory_gb(gpu_name);
-
-                        gpu_specs.push(crate::api::types::GpuSpec {
-                            name: gpu_name.to_string(),
-                            memory_gb,
-                            compute_capability: "8.0".to_string(),
-                        });
-                    }
-                }
-            }
-
-            // Get hardware profile data from joined tables
-            let hw_cpu_model: Option<String> = row.get("cpu_model");
-            let hw_cpu_cores: Option<i32> = row.get("cpu_cores");
-            let hw_ram_gb: Option<i32> = row.get("ram_gb");
-
-            // Get network profile data for location
-            let net_city: Option<String> = row.get("city");
-            let net_region: Option<String> = row.get("region");
-            let net_country: Option<String> = row.get("country");
-
-            // Get speed test data
-            let download_mbps: Option<f64> = row.get("download_mbps");
-            let upload_mbps: Option<f64> = row.get("upload_mbps");
-            let test_timestamp: Option<String> = row.get("test_timestamp");
-
-            // Parse CPU specs from hardware profile data
-            let cpu_specs: crate::api::types::CpuSpec = crate::api::types::CpuSpec {
-                cores: hw_cpu_cores.unwrap_or(0) as u32,
-                model: hw_cpu_model.unwrap_or_else(|| "Unknown".to_string()),
-                memory_gb: hw_ram_gb.unwrap_or(0) as u32,
-            };
-
-            // Build location string from network profile if available
-            let final_location =
-                if net_city.is_some() || net_region.is_some() || net_country.is_some() {
-                    let loc_profile = basilica_common::LocationProfile {
-                        city: net_city,
-                        region: net_region,
-                        country: net_country,
-                    };
-                    Some(loc_profile.to_string())
-                } else {
-                    None
-                };
-
-            // Build network speed info if speed test data is available
-            let network_speed = if download_mbps.is_some() || upload_mbps.is_some() {
-                Some(crate::api::types::NetworkSpeedInfo {
-                    download_mbps,
-                    upload_mbps,
-                    test_timestamp: test_timestamp.and_then(|ts| {
-                        DateTime::parse_from_rfc3339(&ts)
-                            .ok()
-                            .map(|dt| dt.with_timezone(&Utc))
-                    }),
-                })
-            } else {
-                None
-            };
-
-            Ok(Some(crate::api::types::NodeDetails {
-                id: node_id,
-                gpu_specs,
-                cpu_specs,
-                location: final_location,
-                network_speed,
-            }))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Get the actual gpu_count for an node from gpu_uuid_assignments
-    pub async fn get_node_gpu_count_from_assignments(
-        &self,
-        miner_id: &str,
-        node_id: &str,
-    ) -> Result<u32, anyhow::Error> {
-        let count: i64 = sqlx::query_scalar(
-            "SELECT COUNT(DISTINCT gpu_uuid) FROM gpu_uuid_assignments
-             WHERE miner_id = ? AND node_id = ?",
-        )
-        .bind(miner_id)
-        .bind(node_id)
-        .fetch_one(&self.pool)
-        .await?;
-
-        Ok(count as u32)
-    }
-
-    /// Get the actual gpu_memory_gb for a specific GPU index of an node from gpu_uuid_assignments
-    pub async fn get_node_gpu_memory_gb_by_index(
-        &self,
-        miner_id: &str,
-        node_id: &str,
-        index: u32,
-    ) -> Result<f64, anyhow::Error> {
-        let memory: f64 = sqlx::query_scalar(
-            "SELECT COALESCE(gpu_memory_gb, 0.0) FROM gpu_uuid_assignments
-             WHERE miner_id = ? AND node_id = ? AND gpu_index = ?",
-        )
-        .bind(miner_id)
-        .bind(node_id)
-        .bind(index)
-        .fetch_one(&self.pool)
-        .await?;
-
-        Ok(memory)
-    }
-
-    /// Get the actual gpu_memory_gb for a specific GPU index of an node from gpu_uuid_assignments
-    pub async fn get_node_gpu_memory_gb_by_gpu_uuid(
-        &self,
-        miner_id: &str,
-        node_id: &str,
-        gpu_uuid: &str,
-    ) -> Result<f64, anyhow::Error> {
-        let memory: f64 = sqlx::query_scalar(
-            "SELECT COALESCE(gpu_memory_gb, 0.0) FROM gpu_uuid_assignments
-             WHERE miner_id = ? AND node_id = ? AND gpu_uuid = ?",
-        )
-        .bind(miner_id)
-        .bind(node_id)
-        .bind(gpu_uuid)
-        .fetch_one(&self.pool)
-        .await?;
-
-        Ok(memory)
-    }
-
-    /// Get the actual gpu_memory_gb for the first GPU (index 0) of an node from gpu_uuid_assignments
-    pub async fn get_node_first_gpu_memory_gb(
-        &self,
-        miner_id: &str,
-        node_id: &str,
-    ) -> Result<f64, anyhow::Error> {
-        let memory: f64 = sqlx::query_scalar(
-            "SELECT COALESCE(gpu_memory_gb, 0.0) FROM gpu_uuid_assignments
-             WHERE miner_id = ? AND node_id = ? AND gpu_index = 0",
-        )
-        .bind(miner_id)
-        .bind(node_id)
-        .fetch_one(&self.pool)
-        .await?;
-
-        Ok(memory)
-    }
-
-    /// Get the GPU name/model for an node from gpu_uuid_assignments
-    pub async fn get_node_gpu_name_from_assignments(
-        &self,
-        miner_id: &str,
-        node_id: &str,
-    ) -> Result<Option<String>, anyhow::Error> {
-        let gpu_name: Option<String> = sqlx::query_scalar(
-            "SELECT gpu_name FROM gpu_uuid_assignments
-             WHERE miner_id = ? AND node_id = ?
-             LIMIT 1",
-        )
-        .bind(miner_id)
-        .bind(node_id)
-        .fetch_optional(&self.pool)
-        .await?;
-
-        Ok(gpu_name)
-    }
-
-    /// Get the actual gpu_count for all ONLINE nodes of a miner from gpu_uuid_assignments
-    pub async fn get_miner_gpu_uuid_assignments(
-        &self,
-        miner_id: &str,
-    ) -> Result<Vec<(String, u32, String, f64)>, anyhow::Error> {
-        let rows = sqlx::query(
-            "SELECT
-                ga.node_id,
-                COUNT(DISTINCT ga.gpu_uuid) as gpu_count,
-                ga.gpu_name,
-                MAX(ga.gpu_memory_gb) as gpu_memory_gb
-             FROM gpu_uuid_assignments ga
-             JOIN miner_nodes me ON ga.node_id = me.node_id AND ga.miner_id = me.miner_id
-             WHERE ga.miner_id = ?
-                AND me.status IN ('online', 'verified')
-             GROUP BY ga.node_id, ga.gpu_name
-             HAVING COUNT(DISTINCT ga.gpu_uuid) > 0",
-        )
-        .bind(miner_id)
-        .fetch_all(&self.pool)
-        .await?;
-
-        let mut results = Vec::new();
-        for row in rows {
-            let node_id: String = row.get("node_id");
-            let gpu_count: i64 = row.get("gpu_count");
-            let gpu_name: String = row.get("gpu_name");
-            let gpu_memory_gb: f64 = row.get("gpu_memory_gb");
-
-            results.push((node_id, gpu_count as u32, gpu_name, gpu_memory_gb));
-        }
-
-        Ok(results)
-    }
-
-    /// Get total GPU count for a miner from gpu_uuid_assignments
-    pub async fn get_miner_total_gpu_count_from_assignments(
-        &self,
-        miner_id: &str,
-    ) -> Result<u32, anyhow::Error> {
-        let count: i64 = sqlx::query_scalar(
-            "SELECT COUNT(DISTINCT ga.gpu_uuid)
-             FROM gpu_uuid_assignments ga
-             INNER JOIN miner_nodes me ON ga.node_id = me.node_id AND ga.miner_id = me.miner_id
-             WHERE ga.miner_id = ?
-                AND me.status IN ('online', 'verified')",
-        )
-        .bind(miner_id)
-        .fetch_one(&self.pool)
-        .await?;
-
-        Ok(count as u32)
-    }
-
-    /// Get last successful full validation data for lightweight validation
-    pub async fn get_last_full_validation_data(
-        &self,
-        node_id: &str,
-        miner_id: &str,
-    ) -> Result<
-        Option<(
-            f64,
-            Option<super::super::miner_prover::types::NodeResult>,
-            u64,
-            bool,
-        )>,
-        anyhow::Error,
-    > {
-        // duality due to schema migration issues
-        let composite_node_id = if miner_id.starts_with("miner_") {
-            format!("{}__{}", miner_id.replacen("miner_", "miner", 1), node_id)
-        } else {
-            format!("miner{}__{}", miner_id, node_id)
-        };
-
-        let query = r#"
-            SELECT score, details
-            FROM verification_logs
-            WHERE (node_id = ? OR node_id GLOB ('*__' || ?) OR node_id = ? )
-              AND success = 1
-              AND verification_type = 'ssh_automation'
-              AND (
-                json_extract(details, '$.binary_validation_successful') = 1
-                OR json_extract(details, '$.binary_validation_successful') = 'true'
-              )
-            ORDER BY timestamp DESC
-            LIMIT 1
-        "#;
-
-        let row = sqlx::query(query)
-            .bind(node_id)
-            .bind(node_id)
-            .bind(&composite_node_id)
-            .fetch_optional(&self.pool)
-            .await?;
-
-        if let Some(row) = row {
-            let score: f64 = row.get("score");
-            let details_str: String = row.get("details");
-
-            let details: serde_json::Value = serde_json::from_str(&details_str)
-                .map_err(|e| anyhow::anyhow!("Failed to parse details JSON: {}", e))?;
-
-            let node_result = details.get("node_result").and_then(|v| {
-                if v.is_null() {
-                    None
-                } else {
-                    serde_json::from_value::<super::super::miner_prover::types::NodeResult>(
-                        v.clone(),
-                    )
-                    .ok()
-                }
-            });
-
-            let gpu_count = details
-                .get("gpu_count")
-                .and_then(|v| v.as_u64())
-                .unwrap_or(0);
-
-            let binary_validation_successful = details
-                .get("binary_validation_successful")
-                .and_then(|v| v.as_bool())
-                .unwrap_or(false);
-
-            Ok(Some((
-                score,
-                node_result,
-                gpu_count,
-                binary_validation_successful,
-            )))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Get verification count for a miner from recent successful verifications
-    pub async fn get_miner_verification_count(
-        &self,
-        miner_id: &str,
-        hours: i64,
-    ) -> Result<u32, anyhow::Error> {
-        let count_query = r#"
-            SELECT COUNT(*) as count
-            FROM verification_logs vl
-            INNER JOIN miner_nodes me ON vl.node_id = me.node_id
-            WHERE me.miner_id = ?
-            AND vl.success = 1
-            AND vl.timestamp > datetime('now', ? || ' hours')
-        "#;
-
-        let count: i64 = sqlx::query_scalar(count_query)
-            .bind(miner_id)
-            .bind(format!("-{}", hours))
-            .fetch_one(&self.pool)
-            .await
-            .unwrap_or(0);
-
-        Ok(count as u32)
-    }
-
-    /// Get known nodes from database for a miner
-    pub async fn get_known_nodes_for_miner(
-        &self,
-        miner_uid: u16,
-    ) -> Result<Vec<(String, String, i32, String)>, anyhow::Error> {
-        let miner_id = format!("miner_{}", miner_uid);
-
-        let query = r#"
-            SELECT node_id, ssh_endpoint, gpu_count, status
-            FROM miner_nodes
-            WHERE miner_id = ?
-            AND status IN ('online', 'verified')
-            AND (last_health_check IS NULL OR last_health_check > datetime('now', '-1 hour'))
-        "#;
-
-        let rows = sqlx::query(query)
-            .bind(&miner_id)
-            .fetch_all(&self.pool)
-            .await?;
-
-        let mut known_nodes = Vec::new();
-        for row in rows {
-            let node_id: String = row.get("node_id");
-            let ssh_endpoint: String = row.get("ssh_endpoint");
-            let gpu_count: i32 = row.get("gpu_count");
-            let status: String = row.get("status");
-            known_nodes.push((node_id, ssh_endpoint, gpu_count, status));
-        }
-
-        Ok(known_nodes)
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    /// Store node hardware profile information
-    #[allow(clippy::too_many_arguments)]
-    pub async fn store_node_hardware_profile(
-        &self,
-        miner_uid: u16,
-        node_id: &str,
-        cpu_model: Option<String>,
-        cpu_cores: Option<i32>,
-        ram_gb: Option<i32>,
-        disk_gb: Option<i32>,
-        full_hardware_json: &str,
-    ) -> Result<(), anyhow::Error> {
-        sqlx::query(
-            r#"
-            INSERT INTO node_hardware_profile
-            (miner_uid, node_id, cpu_model, cpu_cores, ram_gb, disk_gb, full_hardware_json, updated_at)
-            VALUES (?, ?, ?, ?, ?, ?, ?, CURRENT_TIMESTAMP)
-            ON CONFLICT(miner_uid, node_id) DO UPDATE SET
-                cpu_model = excluded.cpu_model,
-                cpu_cores = excluded.cpu_cores,
-                ram_gb = excluded.ram_gb,
-                disk_gb = excluded.disk_gb,
-                full_hardware_json = excluded.full_hardware_json,
-                updated_at = CURRENT_TIMESTAMP
-            "#,
-        )
-        .bind(miner_uid as i32)
-        .bind(node_id)
-        .bind(cpu_model)
-        .bind(cpu_cores)
-        .bind(ram_gb)
-        .bind(disk_gb)
-        .bind(full_hardware_json)
-        .execute(&self.pool)
-        .await?;
-
-        info!(
-            miner_uid = miner_uid,
-            node_id = node_id,
-            "Stored hardware profile for node"
-        );
-
-        Ok(())
-    }
-
-    /// Retrieve node hardware profile from database
-    pub async fn get_node_hardware_profile(
-        &self,
-        miner_uid: u16,
-        node_id: &str,
-    ) -> Result<
-        Option<(
-            String,
-            Option<String>,
-            Option<i32>,
-            Option<i32>,
-            Option<i32>,
-        )>,
-        anyhow::Error,
-    > {
-        let row = sqlx::query(
-            r#"
-            SELECT cpu_model, cpu_cores, ram_gb, disk_gb, full_hardware_json
-            FROM node_hardware_profile
-            WHERE miner_uid = ? AND node_id = ?
-            "#,
-        )
-        .bind(miner_uid as i32)
-        .bind(node_id)
-        .fetch_optional(&self.pool)
-        .await?;
-
-        if let Some(row) = row {
-            let full_hardware_json: String = row.get("full_hardware_json");
-            let cpu_model: Option<String> = row.get("cpu_model");
-            let cpu_cores: Option<i32> = row.get("cpu_cores");
-            let ram_gb: Option<i32> = row.get("ram_gb");
-            let disk_gb: Option<i32> = row.get("disk_gb");
-
-            Ok(Some((
-                full_hardware_json,
-                cpu_model,
-                cpu_cores,
-                ram_gb,
-                disk_gb,
-            )))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Store node network speedtest profile information
-    #[allow(clippy::too_many_arguments)]
-    pub async fn store_node_speedtest_profile(
-        &self,
-        miner_uid: u16,
-        node_id: &str,
-        download_mbps: Option<f64>,
-        upload_mbps: Option<f64>,
-        test_timestamp: &str,
-        test_server: Option<String>,
-        full_result_json: &str,
-    ) -> Result<(), anyhow::Error> {
-        sqlx::query(
-            r#"
-            INSERT INTO node_speedtest_profile
-            (miner_uid, node_id, download_mbps, upload_mbps, test_timestamp, test_server, full_result_json, updated_at)
-            VALUES (?, ?, ?, ?, ?, ?, ?, CURRENT_TIMESTAMP)
-            ON CONFLICT(miner_uid, node_id) DO UPDATE SET
-                download_mbps = excluded.download_mbps,
-                upload_mbps = excluded.upload_mbps,
-                test_timestamp = excluded.test_timestamp,
-                test_server = excluded.test_server,
-                full_result_json = excluded.full_result_json,
-                updated_at = CURRENT_TIMESTAMP
-            "#,
-        )
-        .bind(miner_uid as i32)
-        .bind(node_id)
-        .bind(download_mbps)
-        .bind(upload_mbps)
-        .bind(test_timestamp)
-        .bind(test_server)
-        .bind(full_result_json)
-        .execute(&self.pool)
-        .await?;
-
-        info!(
-            miner_uid = miner_uid,
-            node_id = node_id,
-            download_mbps = download_mbps.unwrap_or(0.0),
-            upload_mbps = upload_mbps.unwrap_or(0.0),
-            "Stored speedtest profile for node"
-        );
-
-        Ok(())
-    }
-
-    /// Retrieve node speedtest profile from database
-    pub async fn get_node_speedtest_profile(
-        &self,
-        miner_uid: u16,
-        node_id: &str,
-    ) -> Result<Option<(String, Option<f64>, Option<f64>, String, Option<String>)>, anyhow::Error>
-    {
-        let row = sqlx::query(
-            r#"
-            SELECT download_mbps, upload_mbps, test_timestamp, test_server, full_result_json
-            FROM node_speedtest_profile
-            WHERE miner_uid = ? AND node_id = ?
-            "#,
-        )
-        .bind(miner_uid as i32)
-        .bind(node_id)
-        .fetch_optional(&self.pool)
-        .await?;
-
-        if let Some(row) = row {
-            let full_result_json: String = row.get("full_result_json");
-            let download_mbps: Option<f64> = row.get("download_mbps");
-            let upload_mbps: Option<f64> = row.get("upload_mbps");
-            let test_timestamp: String = row.get("test_timestamp");
-            let test_server: Option<String> = row.get("test_server");
-
-            Ok(Some((
-                full_result_json,
-                download_mbps,
-                upload_mbps,
-                test_timestamp,
-                test_server,
-            )))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Store node network geolocation profile information
-    #[allow(clippy::too_many_arguments)]
-    pub async fn store_node_network_profile(
-        &self,
-        miner_uid: u16,
-        node_id: &str,
-        ip_address: Option<String>,
-        hostname: Option<String>,
-        city: Option<String>,
-        region: Option<String>,
-        country: Option<String>,
-        location: Option<String>,
-        organization: Option<String>,
-        postal_code: Option<String>,
-        timezone: Option<String>,
-        test_timestamp: &str,
-        full_result_json: &str,
-    ) -> Result<(), anyhow::Error> {
-        sqlx::query(
-            r#"
-            INSERT INTO node_network_profile
-            (miner_uid, node_id, ip_address, hostname, city, region, country, location,
-             organization, postal_code, timezone, test_timestamp, full_result_json, updated_at)
-            VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, CURRENT_TIMESTAMP)
-            ON CONFLICT(miner_uid, node_id) DO UPDATE SET
-                ip_address = excluded.ip_address,
-                hostname = excluded.hostname,
-                city = excluded.city,
-                region = excluded.region,
-                country = excluded.country,
-                location = excluded.location,
-                organization = excluded.organization,
-                postal_code = excluded.postal_code,
-                timezone = excluded.timezone,
-                test_timestamp = excluded.test_timestamp,
-                full_result_json = excluded.full_result_json,
-                updated_at = CURRENT_TIMESTAMP
-            "#,
-        )
-        .bind(miner_uid as i32)
-        .bind(node_id)
-        .bind(ip_address.clone())
-        .bind(hostname)
-        .bind(city.clone())
-        .bind(region.clone())
-        .bind(country.clone())
-        .bind(location.clone())
-        .bind(organization.clone())
-        .bind(postal_code)
-        .bind(timezone)
-        .bind(test_timestamp)
-        .bind(full_result_json)
-        .execute(&self.pool)
-        .await?;
-
-        info!(
-            miner_uid = miner_uid,
-            node_id = node_id,
-            ip = ip_address.unwrap_or_else(|| "Unknown".to_string()),
-            country = country.unwrap_or_else(|| "Unknown".to_string()),
-            city = city.unwrap_or_else(|| "Unknown".to_string()),
-            region = region.unwrap_or_else(|| "Unknown".to_string()),
-            organization = organization.unwrap_or_else(|| "Unknown".to_string()),
-            location = location.unwrap_or_else(|| "Unknown".to_string()),
-            "Stored network profile for node"
-        );
-
-        Ok(())
-    }
-
-    /// Retrieve node network profile from database
-    #[allow(clippy::type_complexity)]
-    pub async fn get_node_network_profile(
-        &self,
-        miner_uid: u16,
-        node_id: &str,
-    ) -> Result<
-        Option<(
-            String,
-            Option<String>,
-            Option<String>,
-            Option<String>,
-            Option<String>,
-            Option<String>,
-            Option<String>,
-            Option<String>,
-            Option<String>,
-            Option<String>,
-            String,
-        )>,
-        anyhow::Error,
-    > {
-        let row = sqlx::query(
-            r#"
-            SELECT ip_address, hostname, city, region, country, location, organization,
-                   postal_code, timezone, test_timestamp, full_result_json
-            FROM node_network_profile
-            WHERE miner_uid = ? AND node_id = ?
-            "#,
-        )
-        .bind(miner_uid as i32)
-        .bind(node_id)
-        .fetch_optional(&self.pool)
-        .await?;
-
-        if let Some(row) = row {
-            let full_result_json: String = row.get("full_result_json");
-            let ip_address: Option<String> = row.get("ip_address");
-            let hostname: Option<String> = row.get("hostname");
-            let city: Option<String> = row.get("city");
-            let region: Option<String> = row.get("region");
-            let country: Option<String> = row.get("country");
-            let location: Option<String> = row.get("location");
-            let organization: Option<String> = row.get("organization");
-            let postal_code: Option<String> = row.get("postal_code");
-            let timezone: Option<String> = row.get("timezone");
-            let test_timestamp: String = row.get("test_timestamp");
-
-            Ok(Some((
-                full_result_json,
-                ip_address,
-                hostname,
-                city,
-                region,
-                country,
-                location,
-                organization,
-                postal_code,
-                timezone,
-                test_timestamp,
-            )))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Store node Docker validation profile information
-    #[allow(clippy::too_many_arguments)]
-    pub async fn store_node_docker_profile(
-        &self,
-        miner_uid: u16,
-        node_id: &str,
-        service_active: bool,
-        docker_version: Option<String>,
-        images_pulled: Vec<String>,
-        dind_supported: bool,
-        validation_error: Option<String>,
-        full_json: &str,
-    ) -> Result<(), anyhow::Error> {
-        let images_json = serde_json::to_string(&images_pulled)?;
-
-        sqlx::query(
-            r#"
-            INSERT INTO node_docker_profile
-            (miner_uid, node_id, service_active, docker_version, images_pulled,
-             dind_supported, validation_error, full_json, updated_at)
-            VALUES (?, ?, ?, ?, ?, ?, ?, ?, CURRENT_TIMESTAMP)
-            ON CONFLICT(miner_uid, node_id) DO UPDATE SET
-                service_active = excluded.service_active,
-                docker_version = excluded.docker_version,
-                images_pulled = excluded.images_pulled,
-                dind_supported = excluded.dind_supported,
-                validation_error = excluded.validation_error,
-                full_json = excluded.full_json,
-                updated_at = CURRENT_TIMESTAMP
-            "#,
-        )
-        .bind(miner_uid as i32)
-        .bind(node_id)
-        .bind(service_active)
-        .bind(docker_version)
-        .bind(&images_json)
-        .bind(dind_supported)
-        .bind(validation_error)
-        .bind(full_json)
-        .execute(&self.pool)
-        .await?;
-
-        Ok(())
-    }
-
-    /// Get node Docker validation profile
-    pub async fn get_node_docker_profile(
-        &self,
-        miner_uid: u16,
-        node_id: &str,
-    ) -> Result<
-        Option<(
-            String,
-            bool,
-            Option<String>,
-            Vec<String>,
-            bool,
-            Option<String>,
-        )>,
-        anyhow::Error,
-    > {
-        let row = sqlx::query(
-            r#"
-            SELECT service_active, docker_version, images_pulled, dind_supported, validation_error, full_json
-            FROM node_docker_profile
-            WHERE miner_uid = ? AND node_id = ?
-            "#,
-        )
-        .bind(miner_uid as i32)
-        .bind(node_id)
-        .fetch_optional(&self.pool)
-        .await?;
-
-        if let Some(row) = row {
-            let full_json: String = row.get("full_json");
-            let service_active: bool = row.get("service_active");
-            let docker_version: Option<String> = row.get("docker_version");
-            let images_pulled_json: String = row.get("images_pulled");
-            let images_pulled: Vec<String> = serde_json::from_str(&images_pulled_json)?;
-            let dind_supported: bool = row.get("dind_supported");
-            let validation_error: Option<String> = row.get("validation_error");
-
-            Ok(Some((
-                full_json,
-                service_active,
-                docker_version,
-                images_pulled,
-                dind_supported,
-                validation_error,
-            )))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Store node NAT validation profile information
-    pub async fn store_node_nat_profile(
-        &self,
-        miner_uid: u16,
-        node_id: &str,
-        profile: &crate::miner_prover::validation_nat::NatProfile,
-    ) -> Result<(), anyhow::Error> {
-        sqlx::query(
-            r#"
-            INSERT INTO node_nat_profile
-            (miner_uid, node_id, is_accessible, test_port, test_path, container_id,
-             response_content, test_timestamp, full_json, error_message, updated_at)
-            VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, CURRENT_TIMESTAMP)
-            ON CONFLICT(miner_uid, node_id) DO UPDATE SET
-                is_accessible = excluded.is_accessible,
-                test_port = excluded.test_port,
-                test_path = excluded.test_path,
-                container_id = excluded.container_id,
-                response_content = excluded.response_content,
-                test_timestamp = excluded.test_timestamp,
-                full_json = excluded.full_json,
-                error_message = excluded.error_message,
-                updated_at = CURRENT_TIMESTAMP
-            "#,
-        )
-        .bind(miner_uid as i32)
-        .bind(node_id)
-        .bind(profile.is_accessible)
-        .bind(profile.test_port as i32)
-        .bind(&profile.test_path)
-        .bind(&profile.container_id)
-        .bind(&profile.response_content)
-        .bind(profile.test_timestamp.to_rfc3339())
-        .bind(&profile.full_json)
-        .bind(&profile.error_message)
-        .execute(&self.pool)
-        .await?;
-
-        Ok(())
-    }
-
-    /// Get node NAT validation profile
-    pub async fn get_node_nat_profile(
-        &self,
-        miner_uid: u16,
-        node_id: &str,
-    ) -> Result<Option<crate::miner_prover::validation_nat::NatProfile>, anyhow::Error> {
-        let row = sqlx::query(
-            r#"
-            SELECT is_accessible, test_port, test_path, container_id, response_content,
-                   test_timestamp, full_json, error_message
-            FROM node_nat_profile
-            WHERE miner_uid = ? AND node_id = ?
-            "#,
-        )
-        .bind(miner_uid as i32)
-        .bind(node_id)
-        .fetch_optional(&self.pool)
-        .await?;
-
-        if let Some(row) = row {
-            let is_accessible: bool = row.get("is_accessible");
-            let test_port: i32 = row.get("test_port");
-            let test_path: String = row.get("test_path");
-            let container_id: Option<String> = row.get("container_id");
-            let response_content: Option<String> = row.get("response_content");
-            let test_timestamp_str: String = row.get("test_timestamp");
-            let full_json: String = row.get("full_json");
-            let error_message: Option<String> = row.get("error_message");
-
-            let test_timestamp = chrono::DateTime::parse_from_rfc3339(&test_timestamp_str)?
-                .with_timezone(&chrono::Utc);
-
-            Ok(Some(crate::miner_prover::validation_nat::NatProfile {
-                is_accessible,
-                test_port: test_port as u16,
-                test_path,
-                container_id,
-                response_content,
-                test_timestamp,
-                full_json,
-                error_message,
-            }))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Store node storage validation profile information
-    pub async fn store_node_storage_profile(
-        &self,
-        miner_uid: u16,
-        node_id: &str,
-        profile: &crate::miner_prover::validation_storage::StorageProfile,
-    ) -> Result<(), anyhow::Error> {
-        let filesystem_details_json = serde_json::to_string(&profile.filesystem_details)?;
-
-        sqlx::query(
-            r#"
-            INSERT INTO node_storage_profile
-            (miner_uid, node_id, total_bytes, available_bytes,
-             required_bytes, filesystem_details, collection_timestamp, full_json, updated_at)
-            VALUES (?, ?, ?, ?, ?, ?, ?, ?, CURRENT_TIMESTAMP)
-            ON CONFLICT(miner_uid, node_id) DO UPDATE SET
-                total_bytes = excluded.total_bytes,
-                available_bytes = excluded.available_bytes,
-                required_bytes = excluded.required_bytes,
-                filesystem_details = excluded.filesystem_details,
-                collection_timestamp = excluded.collection_timestamp,
-                full_json = excluded.full_json,
-                updated_at = CURRENT_TIMESTAMP
-            "#,
-        )
-        .bind(miner_uid as i32)
-        .bind(node_id)
-        .bind(profile.total_bytes as i64)
-        .bind(profile.available_bytes as i64)
-        .bind(profile.required_bytes as i64)
-        .bind(&filesystem_details_json)
-        .bind(profile.collection_timestamp.to_rfc3339())
-        .bind(&profile.full_json)
-        .execute(&self.pool)
-        .await?;
-
-        Ok(())
-    }
-
-    /// Get node storage validation profile
-    pub async fn get_node_storage_profile(
-        &self,
-        miner_uid: u16,
-        node_id: &str,
-    ) -> Result<Option<crate::miner_prover::validation_storage::StorageProfile>, anyhow::Error>
-    {
-        let row = sqlx::query(
-            r#"
-            SELECT total_bytes, available_bytes, required_bytes,
-                   filesystem_details, collection_timestamp, full_json
-            FROM node_storage_profile
-            WHERE miner_uid = ? AND node_id = ?
-            "#,
-        )
-        .bind(miner_uid as i32)
-        .bind(node_id)
-        .fetch_optional(&self.pool)
-        .await?;
-
-        if let Some(row) = row {
-            let total_bytes: i64 = row.get("total_bytes");
-            let available_bytes: i64 = row.get("available_bytes");
-            let required_bytes: i64 = row.get("required_bytes");
-            let filesystem_details_json: String = row.get("filesystem_details");
-            let collection_timestamp_str: String = row.get("collection_timestamp");
-            let full_json: String = row.get("full_json");
-
-            let filesystem_details = serde_json::from_str(&filesystem_details_json)?;
-            let collection_timestamp =
-                chrono::DateTime::parse_from_rfc3339(&collection_timestamp_str)?
-                    .with_timezone(&chrono::Utc);
-
-            Ok(Some(
-                crate::miner_prover::validation_storage::StorageProfile {
-                    total_bytes: total_bytes as u64,
-                    available_bytes: available_bytes as u64,
-                    required_bytes: required_bytes as u64,
-                    filesystem_details,
-                    collection_timestamp,
-                    full_json,
-                },
-            ))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Get all nodes with their GPU and rental data for metrics initialization
-    /// This eliminates N+1 queries by fetching everything in a single query with joins
-    pub async fn get_all_nodes_for_metrics(&self) -> Result<Vec<NodeMetricData>, anyhow::Error> {
-        let query = r#"
-            SELECT
-                me.node_id,
-                me.miner_id,
-                gua.gpu_name,
-                CASE WHEN r.id IS NOT NULL THEN 1 ELSE 0 END as has_active_rental
-            FROM miner_nodes me
-            INNER JOIN gpu_uuid_assignments gua
-                ON me.node_id = gua.node_id
-                AND me.miner_id = gua.miner_id
-            LEFT JOIN rentals r
-                ON me.node_id = r.node_id
-                AND me.miner_id = r.miner_id
-                AND r.state = 'active'
-            WHERE gua.gpu_name IS NOT NULL
-            GROUP BY me.node_id, me.miner_id
-        "#;
-
-        let rows = sqlx::query(query).fetch_all(&self.pool).await?;
-
-        let mut node_metrics = Vec::new();
-        for row in rows {
-            let node_id: String = row.get("node_id");
-            let miner_id: String = row.get("miner_id");
-            let gpu_name: Option<String> = row.get("gpu_name");
-            let has_active_rental: i32 = row.get("has_active_rental");
-
-            // Extract miner UID from miner_id string (format: "miner_{uid}")
-            let miner_uid = if miner_id.starts_with("miner_") {
-                miner_id
-                    .strip_prefix("miner_")
-                    .and_then(|uid_str| uid_str.parse::<u16>().ok())
-                    .unwrap_or(0)
-            } else {
-                0
-            };
-
-            node_metrics.push(NodeMetricData {
-                node_id,
-                miner_id,
-                miner_uid,
-                gpu_name,
-                has_active_rental: has_active_rental != 0,
-            });
-        }
-
-        Ok(node_metrics)
     }
 }
 
@@ -2428,102 +192,6 @@ impl ValidatorPersistence for SimplePersistence {
     }
 }
 
-/// Nodestatistics derived from verification logs
-#[derive(Debug, Clone)]
-pub struct NodeStats {
-    pub node_id: String,
-    pub total_verifications: u64,
-    pub successful_verifications: u64,
-    pub average_score: Option<f64>,
-    pub average_duration_ms: Option<f64>,
-    pub first_verification: Option<DateTime<Utc>>,
-    pub last_verification: Option<DateTime<Utc>>,
-}
-
-impl NodeStats {
-    pub fn success_rate(&self) -> f64 {
-        if self.total_verifications == 0 {
-            0.0
-        } else {
-            self.successful_verifications as f64 / self.total_verifications as f64
-        }
-    }
-}
-
-/// Available capacity entry
-#[derive(Debug, Clone)]
-pub struct CapacityEntry {
-    pub node_id: String,
-    pub verification_score: f64,
-    pub success_rate: f64,
-    pub last_verification: DateTime<Utc>,
-    pub hardware_info: Value,
-    pub total_verifications: u64,
-}
-
-/// Miner data for listings
-#[derive(Debug, Clone)]
-pub struct MinerData {
-    pub miner_id: String,
-    pub hotkey: String,
-    pub endpoint: String,
-    pub node_count: u32,
-    pub verification_score: f64,
-    pub uptime_percentage: f64,
-    pub last_seen: DateTime<Utc>,
-    pub registered_at: DateTime<Utc>,
-    pub node_info: Value,
-}
-
-/// Miner health data
-#[derive(Debug, Clone)]
-pub struct MinerHealthData {
-    pub last_health_check: DateTime<Utc>,
-    pub node_health: Vec<NodeHealthData>,
-}
-
-#[derive(Debug, Clone)]
-pub struct NodeHealthData {
-    pub node_id: String,
-    pub status: String,
-    pub last_seen: DateTime<Utc>,
-}
-
-/// Nodedetails for miner listings
-#[derive(Debug, Clone)]
-pub struct NodeData {
-    pub node_id: String,
-    pub gpu_specs: Vec<crate::api::types::GpuSpec>,
-    pub cpu_specs: crate::api::types::CpuSpec,
-    pub location: Option<String>,
-}
-
-/// Available node data for rental listings
-#[derive(Debug, Clone)]
-pub struct AvailableNodeData {
-    pub node_id: String,
-    pub miner_id: String,
-    pub gpu_specs: Vec<crate::api::types::GpuSpec>,
-    pub cpu_specs: crate::api::types::CpuSpec,
-    pub location: Option<String>,
-    pub verification_score: f64,
-    pub uptime_percentage: f64,
-    pub status: Option<String>,
-    pub download_mbps: Option<f64>,
-    pub upload_mbps: Option<f64>,
-    pub speed_test_timestamp: Option<chrono::DateTime<chrono::Utc>>,
-}
-
-/// Node metric data for initializing metrics
-#[derive(Debug, Clone)]
-pub struct NodeMetricData {
-    pub node_id: String,
-    pub miner_id: String,
-    pub miner_uid: u16,
-    pub gpu_name: Option<String>,
-    pub has_active_rental: bool,
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -2535,7 +203,6 @@ mod tests {
             .await
             .expect("Failed to create persistence");
 
-        // First miner registration
         let nodes1 = vec![NodeRegistration {
             node_id: "exec1".to_string(),
             ssh_endpoint: "http://192.168.1.1:8080".to_string(),
@@ -2552,16 +219,14 @@ mod tests {
             },
         }];
 
-        // Register first miner successfully
         let result = persistence
             .register_miner("miner1", "hotkey1", "http://miner1.com", &nodes1)
             .await;
         assert!(result.is_ok());
 
-        // Second miner trying to register with same ssh_endpoint
         let nodes2 = vec![NodeRegistration {
             node_id: "exec2".to_string(),
-            ssh_endpoint: "http://192.168.1.1:8080".to_string(), // Same address!
+            ssh_endpoint: "http://192.168.1.1:8080".to_string(),
             gpu_count: 1,
             gpu_specs: vec![GpuSpec {
                 name: "RTX 3090".to_string(),
@@ -2575,7 +240,6 @@ mod tests {
             },
         }];
 
-        // Should fail due to duplicate ssh_endpoint
         let result = persistence
             .register_miner("miner2", "hotkey2", "http://miner2.com", &nodes2)
             .await;
@@ -2592,7 +256,6 @@ mod tests {
             .await
             .expect("Failed to create persistence");
 
-        // Register first miner
         let nodes1 = vec![NodeRegistration {
             node_id: "exec1".to_string(),
             ssh_endpoint: "http://192.168.1.1:8080".to_string(),
@@ -2610,7 +273,6 @@ mod tests {
             .await
             .expect("Failed to register miner1");
 
-        // Register second miner with different address
         let nodes2 = vec![NodeRegistration {
             node_id: "exec2".to_string(),
             ssh_endpoint: "http://192.168.1.2:8080".to_string(),
@@ -2628,12 +290,11 @@ mod tests {
             .await
             .expect("Failed to register miner2");
 
-        // Try to update miner2 with miner1's ssh_endpoint
         let update_request = UpdateMinerRequest {
             endpoint: None,
             nodes: Some(vec![NodeRegistration {
                 node_id: "exec2_updated".to_string(),
-                ssh_endpoint: "http://192.168.1.1:8080".to_string(), // Trying to use miner1's address
+                ssh_endpoint: "http://192.168.1.1:8080".to_string(),
                 gpu_count: 1,
                 gpu_specs: vec![],
                 cpu_specs: CpuSpec {
@@ -2659,7 +320,6 @@ mod tests {
             .await
             .expect("Failed to create persistence");
 
-        // Register miner
         let nodes = vec![NodeRegistration {
             node_id: "exec1".to_string(),
             ssh_endpoint: "http://192.168.1.1:8080".to_string(),
@@ -2677,18 +337,17 @@ mod tests {
             .await
             .expect("Failed to register miner");
 
-        // Update same miner with same ssh_endpoint (should succeed)
         let update_request = UpdateMinerRequest {
             endpoint: Some("http://miner1-updated.com".to_string()),
             nodes: Some(vec![NodeRegistration {
                 node_id: "exec1_updated".to_string(),
-                ssh_endpoint: "http://192.168.1.1:8080".to_string(), // Same address is OK for same miner
-                gpu_count: 3,                                        // Updated GPU count
+                ssh_endpoint: "http://192.168.1.1:8080".to_string(),
+                gpu_count: 3,
                 gpu_specs: vec![],
                 cpu_specs: CpuSpec {
                     cores: 16,
                     model: "Intel i9".to_string(),
-                    memory_gb: 64, // Updated memory
+                    memory_gb: 64,
                 },
             }]),
             signature: "test_signature".to_string(),
@@ -2702,7 +361,6 @@ mod tests {
     async fn test_gpu_uuid_duplicate_prevention() {
         let persistence = SimplePersistence::for_testing().await.unwrap();
 
-        // Register initial miner with node
         let node1 = NodeRegistration {
             node_id: "exec1".to_string(),
             ssh_endpoint: "root@192.168.1.100:50051".to_string(),
@@ -2720,7 +378,6 @@ mod tests {
             .await
             .unwrap();
 
-        // Manually insert GPU UUID for testing
         let gpu_uuid = "GPU-550e8400-e29b-41d4-a716-446655440000";
         sqlx::query("UPDATE miner_nodes SET gpu_uuids = ? WHERE miner_id = ? AND node_id = ?")
             .bind(gpu_uuid)
@@ -2730,7 +387,6 @@ mod tests {
             .await
             .unwrap();
 
-        // Register another miner with different node
         let node2 = NodeRegistration {
             node_id: "exec2".to_string(),
             ssh_endpoint: "root@192.168.1.101:50051".to_string(),
@@ -2748,14 +404,12 @@ mod tests {
             .await
             .unwrap();
 
-        // Verify both nodes exist
         let count: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM miner_nodes")
             .fetch_one(&persistence.pool)
             .await
             .unwrap();
         assert_eq!(count, 2);
 
-        // Verify only one has the GPU UUID
         let gpu_count: i64 =
             sqlx::query_scalar("SELECT COUNT(*) FROM miner_nodes WHERE gpu_uuids = ?")
                 .bind(gpu_uuid)
@@ -2771,7 +425,6 @@ mod tests {
             .await
             .expect("Failed to create persistence");
 
-        // Register a miner with an node
         let node = NodeRegistration {
             node_id: "exec1".to_string(),
             ssh_endpoint: "root@192.168.1.100:50051".to_string(),
@@ -2789,10 +442,9 @@ mod tests {
             .await
             .unwrap();
 
-        // Store hardware profile for the node
         persistence
             .store_node_hardware_profile(
-                1, // miner_uid
+                1,
                 "exec1",
                 Some("AMD EPYC 7763".to_string()),
                 Some(64),
@@ -2803,10 +455,9 @@ mod tests {
             .await
             .unwrap();
 
-        // Store network profile for the node
         persistence
             .store_node_network_profile(
-                1, // miner_uid
+                1,
                 "exec1",
                 Some("192.168.1.100".to_string()),
                 Some("exec1.example.com".to_string()),
@@ -2823,7 +474,6 @@ mod tests {
             .await
             .unwrap();
 
-        // Get miner nodes and verify hardware profile is used
         let nodes = persistence.get_miner_nodes("miner_1").await.unwrap();
         assert_eq!(nodes.len(), 1);
 
@@ -2837,7 +487,6 @@ mod tests {
             Some("San Francisco/California/US".to_string())
         );
 
-        // Test get_available_nodes with hardware profile
         let available = persistence
             .get_available_nodes(None, None, None, None)
             .await
